requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
Pillow>=10.0.0
